+++ src/elements/conditions/entries/TypeConditionRule.php
@@ -3,17 +3,12 @@
 namespace craft\elements\conditions\entries;
 
 use Craft;
-use craft\base\conditions\BaseConditionRule;
+use craft\base\conditions\BaseMultiSelectConditionRule;
 use craft\base\ElementInterface;
-use craft\db\Table;
 use craft\elements\conditions\ElementConditionRuleInterface;
 use craft\elements\db\ElementQueryInterface;
 use craft\elements\db\EntryQuery;
 use craft\elements\Entry;
-use craft\helpers\ArrayHelper;
-use craft\helpers\Cp;
-use craft\helpers\Db;
-use craft\helpers\UrlHelper;
 use craft\models\Section;
 
 /**
@@ -22,7 +17,7 @@ use craft\models\Section;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 4.0.0
  */
-class TypeConditionRule extends BaseConditionRule implements ElementConditionRuleInterface
+class TypeConditionRule extends BaseMultiSelectConditionRule implements ElementConditionRuleInterface
 {
     /**
      * @inheritdoc
@@ -45,7 +40,10 @@ class TypeConditionRule extends BaseConditionRule implements ElementConditionRul
      */
     private array $_sections = [];
 
+    /** @deprecated in 4.0.1.  */
     public ?string $sectionUid = null;
+
+    /** @deprecated in 4.0.1.  */
     public ?string $entryTypeUid = null;
 
     /**
@@ -55,47 +53,18 @@ class TypeConditionRule extends BaseConditionRule implements ElementConditionRul
     {
         $this->_sections = Craft::$app->getSections()->getAllSections();
 
-        // Set a default section
-        if (!$this->sectionUid) {
-            $this->sectionUid = ArrayHelper::firstValue($this->_sections)?->uid;
-        }
-
-        // Once we have a section, set a default entry type
-        $this->_ensureEntryType();
-
         parent::init();
     }
 
     /**
-     * @inheritdoc
-     */
-    public function getConfig(): array
-    {
-        return array_merge(parent::getConfig(), [
-            'sectionUid' => $this->sectionUid,
-            'entryTypeUid' => $this->entryTypeUid,
-        ]);
-    }
-
-    /**
      * @return array
      */
-    private function _sectionOptions(): array
-    {
-        return ArrayHelper::map($this->_sections, 'uid', 'name');
-    }
-
-    /**
-     * @return array
-     */
-    private function _entryTypeOptions(): array
+    protected function options(): array
     {
         $options = [];
         foreach ($this->_sections as $section) {
-            if ($section->uid == $this->sectionUid) {
-                foreach ($section->getEntryTypes() as $entryType) {
-                    $options[$entryType->uid] = $entryType->name;
-                }
+            foreach ($section->getEntryTypes() as $entryType) {
+                $options[$entryType->uid] = sprintf('%s - %s', $section->name, $entryType->name);
             }
         }
 
@@ -105,64 +74,16 @@ class TypeConditionRule extends BaseConditionRule implements ElementConditionRul
     /**
      * @inheritdoc
      */
-    protected function inputHtml(): string
-    {
-        $html = Cp::selectHtml([
-            'name' => 'sectionUid',
-            'value' => $this->sectionUid,
-            'options' => $this->_sectionOptions(),
-            'inputAttributes' => [
-                'hx' => [
-                    'post' => UrlHelper::actionUrl('conditions/render'), // Only the section re-renders the body
-                    'target' => 'closest .rule-body',
-                    'select' => '#' . Craft::$app->getView()->namespaceInputId('rule-body'),
-                    'swap' => 'outerHTML',
-                ],
-            ],
-        ]);
-
-        $this->_ensureEntryType();
-
-        $html .= Cp::selectHtml([
-            'name' => 'entryTypeUid',
-            'value' => $this->entryTypeUid,
-            'options' => $this->_entryTypeOptions(),
-        ]);
-
-        return $html;
-    }
-
-    /**
-     * Ensures an entry type is set correctly based on the section selected.
-     */
-    private function _ensureEntryType(): void
-    {
-        if (!$this->sectionUid) {
-            $this->entryTypeUid = null;
-            return;
-        }
-
-        $entryTypeOptions = $this->_entryTypeOptions();
-
-        if (!$this->entryTypeUid || !ArrayHelper::keyExists($this->entryTypeUid, $entryTypeOptions)) {
-            $this->entryTypeUid = ArrayHelper::firstKey($entryTypeOptions);
-        }
-    }
-
-    /**
-     * @inheritdoc
-     */
     public function modifyQuery(ElementQueryInterface $query): void
     {
-        $typeId = Db::idByUid(Table::ENTRYTYPES, $this->entryTypeUid);
-        $type = Craft::$app->getSections()->getEntryTypeById($typeId);
-
         /** @var EntryQuery $query */
-        $query->type($type);
+        $sections = Craft::$app->getSections();
+        $query->typeId($this->paramValue(fn($uid) => $sections->getEntryTypeByUid($uid)->id ?? null));
     }
 
     /**
      * @inheritdoc
+     * @TODO Remove when `sectionUid` and `entryTypeUid` are removed.
      */
     protected function defineRules(): array
     {
@@ -177,7 +98,6 @@ class TypeConditionRule extends BaseConditionRule implements ElementConditionRul
     public function matchElement(ElementInterface $element): bool
     {
         /** @var Entry $element */
-        $typeId = Db::idByUid(Table::ENTRYTYPES, $this->entryTypeUid);
-        return $element->getType()->id === (int)$typeId;
+        return $this->matchValue((string)$element->getType()->id);
     }
 }