+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 ## Unreleased
 
+### Changed
+- Entry Type condition rules now all the selection of multiple entry types.
+
 ### Fixed
 - Fixed a bug where Money field labels’ `for` attributes weren’t referencing the correct input ID. ([#11016](https://github.com/craftcms/cms/pull/11016))
 - Fixed a bug where Money field inputs weren’t getting `aria-describedby` attributes. ([#11016](https://github.com/craftcms/cms/pull/11016))
 - Fixed an error that occurred when loading an edit screen for an element type that didn’t have a field layout. ([#11110](https://github.com/craftcms/cms/pull/11110))
 - Fixed a bug where condition rules they weren’t selectable (per `isSelectable()`) were still visible in the rule dropdown menu. ([#11104](https://github.com/craftcms/cms/pull/11104))
+
+### Deprecated
+- Deprecated `craft\elements\conditions\entries\EntryTypeCondition::$sectionUid`.
+- Deprecated `craft\elements\conditions\entries\EntryTypeCondition::$entryTypeUid`.
 
 ## 4.0.0.1 - 2022-05-04